backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.1
uvicorn[standard]==0.25.0
gunicorn==21.2.0
pydantic[email]==2.12.5
pydantic-settings==2.1.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
mysql-connector-python==8.2.0

# Authentication & Security
PyJWT==2.10.1
bcrypt==4.1.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-dotenv==1.2.1

# Image Processing & PDF Generation
pillow==10.4.0
reportlab==4.0.9
qrcode==8.0

# Email & Communication
email-validator==2.3.0
twilio==9.9.0

# HTTP Requests
requests==2.32.5
aiohttp==3.13.2

# Development Tools (optional)
black==25.12.0
flake8==7.3.0
isort==7.0.0
mypy==1.19.0
pytest==9.0.2

# Data Processing (if needed)
pandas==2.3.3
numpy==2.3.5

# Other utilities
python-dateutil==2.9.0.post0
pytz==2025.2